pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "roletrando"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "prefect>=3",
    "httpx",
    "adbc-driver-gizmosql",
    "pyarrow",
    "tinytuya",
    "gspread",
    "pandas",
    "google-generativeai",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["flows*"]
